+++ package.json
@@ -2,7 +2,8 @@
   "name": "little-saga",
   "license": "MIT",
   "description": "Just a little-saga.",
-  "version": "0.5.4",
+  "version": "0.6.0-SNAPSHOT",
+  "repository": "https://github.com/little-saga/little-saga.git",
   "scripts": {
     "build": "cross-env BABEL_ENV=es rollup -c rollup.config.js",
     "test": "jest",
