+++ src/defaultScheduler.js
@@ -1,0 +1,3 @@
+import makeScheduler from './makeScheduler'
+
+export default makeScheduler()
