+++ src/index.js
@@ -2,6 +2,7 @@ export * from './symbols'
 export * from './utils'
 export * from './saga-helpers'
 export * from './channels'
+export * from './stdChannel'
 export * from './io-helpers'
 
 import * as buffers from './buffers'
@@ -11,3 +12,4 @@ export { buffers }
 export { default as io } from './io'
 export { default as runSaga } from './runSaga'
 export { default as makeScheduler } from './makeScheduler'
+export { defualt as defaultScheduler } from './defaultScheduler'
