+++ src/io-helpers.js
@@ -1,8 +1,4 @@
-import { IO } from './symbols'
-
-export function makeEffect(type, payload) {
-  return { [IO]: true, type, payload }
-}
+import { makeEffect } from './internal-utils'
 
 export function detach({ type, payload }) {
   return makeEffect(type, { ...payload, detached: true })
